+++ NEWS.md
@@ -3,3 +3,4 @@
 * documentation spell-check
 * the result of algebra() is now an S3 class, implemented print.algebra()
 * added involutive negation to algebra()$ni
+* added nelson(), lowerEst()