+++ NEWS.md
@@ -3,4 +3,5 @@
 * documentation spell-check
 * the result of algebra() is now an S3 class, implemented print.algebra()
 * added involutive negation to algebra()$ni
-* added nelson(), lowerEst()
+* added nelson(), lowerEst() handlers of NA for algebra()
+* added print() method for linguistic contexts