+++ NEWS.md
@@ -1,5 +1,4 @@
 # lfl 2.1.0
-* documentation spell-check
 * the result of algebra() is now an S3 class, implemented print.algebra()
 * added involutive negation to algebra()$ni
 * added nelson(), lowerEst() handlers of NA for algebra()
@@ -8,3 +7,9 @@
 * implemented bi-residuum of the dragonfly algebra
 * implemented order operation of the dragonfly and lowerEst algebras
 * added fuzzy quantifiers
+
+
+# lfl 2.0.1
+* added support for percentual quantifiers in compose()
+* documentation spell-check
+